feat: Add active effects to Twodsix Actors (#1197)
This update adds the ability to add active effects to actors (Travellers) and most items.  Active effects are useful in temporarily modifying actor attributes.  See the FVTT knowledge base for more information (https://foundryvtt.com/article/active-effects/).  For example, powered armor - when worn might - convey a strength bonus.  This is a good use of an active effect.

Note that you must opt-into using active effects for items (in debug menu).  Also, the effect from items to actors is only enabled when the item is equipped (if applicable).  Because this is an complex feature, there is a new entry on the Twodsix wiki (https://github.com/xdy/twodsix-foundryvtt/wiki/Using-Active-Effects-and-Twodsix) that goes into a bit more detail.

As with any major update like this there are bound to be some bugs.  So please don't update right before a game session and also try it on a duplicate world first.

Thanks go to @jonepatr who wrote most of the code.  This update was a lot of work.  Also, thanks to @ForjaSalvaje for testing a beta release and suggesting some great improvements.