feat: add active effects custom mode (#1518)
This update adds a CUSTOM calculation mode for active effects.  With this mode, one can specify the effect value as a formula and not simply a static variable.  Please note that this mode ADDS the effect value to the key value.  A worked example is on the wiki: https://github.com/xdy/twodsix-foundryvtt/wiki/Using-Active-Effects-and-Twodsix#custom-change-mode.  It is possible to create circular references - the effects will be odd for such cases.  Thanks to @Amantiado who raised the issue of making Powered Armor effects work.